Refactor: Move users tracking code to UsersTracking module
